Added feature to disable the Abort button when no sequence is running and only have it clickable when something is being executed
