+++ app/src/components/CardRow.js
@@ -3,11 +3,11 @@ import { Col } from "react-bootstrap";
 
 /* Creates a row within the accordion, for each sequence to go into. */
 
-const CardRow = ({sequences, row_title, executionPanelRef}) => {
+const CardRow = ({sequences, row_title, executionPanelRef, setAbortDisabled}) => {
     sequences = JSON.parse(sequences)
     const listItems = (Object.entries(sequences)).flatMap(([sequenceKey, sequence]) => 
         <Col>
-            <SequenceCard key={sequenceKey} header={sequenceKey} sequence={sequence} row_title={row_title} executionPanelRef={executionPanelRef}></SequenceCard>
+            <SequenceCard key={sequenceKey} header={sequenceKey} sequence={sequence} row_title={row_title} executionPanelRef={executionPanelRef} setAbortDisabled={setAbortDisabled}></SequenceCard>
         </Col>
     );
     return (listItems)
