+++ app/src/components/ModuleList.js
@@ -3,7 +3,7 @@ import Accordion from 'react-bootstrap/Accordion';
 
 /* Initialises a SequenceModule component for each module found in the object. */
 
-const ModuleList = ({sequence_modules, executionPanelRef}) => {
+const ModuleList = ({sequence_modules, executionPanelRef, setAbortDisabled}) => {
 
     sequence_modules = Object.fromEntries(Object.entries(sequence_modules).sort());
 
@@ -11,7 +11,7 @@ const ModuleList = ({sequence_modules, executionPanelRef}) => {
     const sequences_modules_obj = sequence_modules
 
     const listHeaders = Object.entries(sequences_modules_obj).flatMap(([moduleKey,sequences]) =>
-        <SequenceModule key={moduleKey} header={moduleKey} sequences={JSON.stringify(sequences)} executionPanelRef={executionPanelRef}></SequenceModule>
+        <SequenceModule key={moduleKey} header={moduleKey} sequences={JSON.stringify(sequences)} executionPanelRef={executionPanelRef} setAbortDisabled={setAbortDisabled}></SequenceModule>
     )
 
     return (
