+++ app/src/components/SequenceModule.js
@@ -4,14 +4,14 @@ import { Row } from 'react-bootstrap'
 
 /* Constructs Accordion layer for each sequence module, then intialises the CardRow to display sequences. */
 
-const SequenceModule = ({sequences, header, executionPanelRef}) => {
+const SequenceModule = ({sequences, header, executionPanelRef, setAbortDisabled}) => {
     console.log("ACC ITEM", sequences, header)
     return (
         <Accordion.Item eventKey={header}>
             <Accordion.Header>{header}</Accordion.Header>
             <Accordion.Body>
                 <Row>
-                    <CardRow sequences={sequences} row_title={header} executionPanelRef={executionPanelRef}></CardRow>
+                    <CardRow sequences={sequences} row_title={header} executionPanelRef={executionPanelRef} setAbortDisabled={setAbortDisabled}></CardRow>
                 </Row>
             </Accordion.Body>
         </Accordion.Item>
